+++ CHANGELOG.md
@@ -6,13 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- `sample-2`: A sample where all one has to do to compile is run `npm install` and `npm start`
 ### Changed
-- removed debug output from `process-dita.js`
 ### Deprecated
 ### Removed
 ### Fixed
 ### Security
+## [0.10.1] - 2020-06-12
+### Added
+- `sample-2`: A sample where all one has to do to compile is run `npm install` and `npm start`
+### Changed
+- removed debug output from `process-dita.js`
 ## [0.10.0] - 2020-06-12
 ### Added
 - ability to install a temporary DITA-OT instance using the `-i` flag
@@ -63,7 +66,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## v0.9.0 - 2020-06-11
 Initial prerelease.
 
-[Unreleased]: https://github.com/fliegwerk/dita-ot-helper/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/fliegwerk/dita-ot-helper/compare/v0.10.1...HEAD
 [0.9.1]: https://github.com/fliegwerk/dita-ot-helper/compare/v0.9.0...v0.9.1
 [0.9.2]: https://github.com/fliegwerk/dita-ot-helper/compare/v0.9.1...v0.9.2
 [0.9.3]: https://github.com/fliegwerk/dita-ot-helper/compare/v0.9.2...v0.9.3
@@ -73,3 +76,4 @@ Initial prerelease.
 [0.9.7]: https://github.com/fliegwerk/dita-ot-helper/compare/v0.9.6...v0.9.7
 [0.9.8]: https://github.com/fliegwerk/dita-ot-helper/compare/v0.9.7...v0.9.8
 [0.10.0]: https://github.com/fliegwerk/dita-ot-helper/compare/v0.9.8...v0.10.0
+[0.10.1]: https://github.com/fliegwerk/dita-ot-helper/compare/v0.10.0...v0.10.1